break cyclic import by moving load_source from tools.filetools to tools.hooks
